Move `io.getquill.CassandraStreamContext` into `quill-cassandra-streaming-monix` module
