+++ quill-cassandra-streaming-monix/src/main/scala/io/getquill/CassandraStreamContext.scala
