+++ quill-cassandra-streaming-monix/src/test/resources/application.conf
@@ -1,0 +1,6 @@
+testStreamDB.keyspace=quill_test
+testStreamDB.preparedStatementCacheSize=1000
+testStreamDB.session.contactPoint=${?CASSANDRA_HOST}
+testStreamDB.session.port=${?CASSANDRA_PORT}
+testStreamDB.session.queryOptions.fetchSize=1
+testStreamDB.session.queryOptions.consistencyLevel=LOCAL_QUORUM
