+++ quill-cassandra-streaming-monix/src/test/scala/io/getquill/context/cassandra/QueryResultTypeCassandraStreamSpec.scala
