+++ quill-cassandra-streaming-monix/src/test/scala/io/getquill/context/cassandra/package.scala
@@ -1,0 +1,11 @@
+package io.getquill.context
+
+import io.getquill.Literal
+
+import io.getquill.CassandraStreamContext
+
+
+package object cassandra {
+
+  lazy val testStreamDB = new CassandraStreamContext(Literal, "testStreamDB") with CassandraTestEntities
+}
