+++ quill-cassandra/src/test/resources/application.conf
@@ -17,10 +17,3 @@ testAsyncDB.session.contactPoint=${?CASSANDRA_HOST}
 testAsyncDB.session.port=${?CASSANDRA_PORT}
 testAsyncDB.session.queryOptions.fetchSize=1
 testAsyncDB.session.queryOptions.consistencyLevel=LOCAL_QUORUM
-
-testStreamDB.keyspace=quill_test
-testStreamDB.preparedStatementCacheSize=1000
-testStreamDB.session.contactPoint=${?CASSANDRA_HOST}
-testStreamDB.session.port=${?CASSANDRA_PORT}
-testStreamDB.session.queryOptions.fetchSize=1
-testStreamDB.session.queryOptions.consistencyLevel=LOCAL_QUORUM
