+++ quill-cassandra/src/test/scala/io/getquill/context/cassandra/DecodeNullSpec.scala
@@ -1,7 +1,6 @@
 package io.getquill.context.cassandra
 
 import io.getquill._
-import monix.reactive.Observable
 
 class DecodeNullSpec extends Spec {
 
@@ -34,26 +33,6 @@ class DecodeNullSpec extends Spec {
       intercept[IllegalStateException] {
         await {
           result
-        }
-      }
-    }
-
-    "stream" in {
-      import testStreamDB._
-      import monix.execution.Scheduler.Implicits.global
-      val writeEntities = quote(querySchema[DecodeNullTestWriteEntity]("DecodeNullTestEntity"))
-
-      val result =
-        for {
-          _ <- testStreamDB.run(writeEntities.delete)
-          _ <- Observable.fromTask(testStreamDB.run(writeEntities.insert(lift(insertValue))).countL)
-          result <- testStreamDB.run(query[DecodeNullTestEntity])
-        } yield {
-          result
-        }
-      intercept[IllegalStateException] {
-        await {
-          result.headL.runAsync
         }
       }
     }
