+++ quill-cassandra/src/test/scala/io/getquill/context/cassandra/package.scala
@@ -6,7 +6,6 @@ import scala.concurrent.duration.Duration
 
 import io.getquill._
 import io.getquill.Literal
-import io.getquill.CassandraStreamContext
 
 package object cassandra {
 
@@ -15,8 +14,6 @@ package object cassandra {
   lazy val testSyncDB = new CassandraSyncContext(Literal, "testSyncDB") with CassandraTestEntities
 
   lazy val testAsyncDB = new CassandraAsyncContext(Literal, "testAsyncDB") with CassandraTestEntities
-
-  lazy val testStreamDB = new CassandraStreamContext(Literal, "testStreamDB") with CassandraTestEntities
 
   def await[T](f: Future[T]): T = Await.result(f, Duration.Inf)
 }
